Restructured script into modules
