+++ index.html
@@ -3,7 +3,7 @@
   <head>
     <title>Fractal Clock</title>
     <link rel="stylesheet" href="styles.css"></link>
-    <script type="module" src="main.js"></script>
+    <script type="module" src="scripts/main.js"></script>
     <link rel="icon" href="icon.png"></link>
     <link rel="preconnect" href="https://fonts.googleapis.com">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
@@ -48,6 +48,6 @@
       </div>
     </div>
     <a class="githubButton" href="https://github.com/BrickNumber5/FractalClock" target="_blank">GitHub</a>
-    <button class="hyperspeedButton aish" onclick="hyperspeed();">Hyperspeed</button>
+    <button class="hyperspeedButton aish" onclick="toggleHyperspeed();">Hyperspeed</button>
   </body>
 </html>
